chatbot-backend/requirements.txt
#pip install -r requirements.txt
fastapi
uvicorn
sqlmodel
pydantic
mysqlclient  # 또는 pymysql
python-dotenv
pytest
requests
openai
passlib[bcrypt]
PyJWT
